Extracted license-related storage code into its own class with its own interface, update the DI configuration.
