+++ src/NuGetGallery.Core/Services/CorePackageFileService.cs
@@ -13,8 +13,6 @@ namespace NuGetGallery
 {
     public class CorePackageFileService : ICorePackageFileService
     {
-        private const string LicenseFileName = "license";
-
         private readonly ICoreFileStorageService _fileStorageService;
         private readonly IFileMetadataService _metadata;
 
@@ -36,25 +34,25 @@ namespace NuGetGallery
                 throw new ArgumentNullException(nameof(packageFile));
             }
 
-            var fileName = BuildFileName(package, _metadata.FileSavePathTemplate, _metadata.FileExtension);
+            var fileName = FileNameHelper.BuildFileName(package, _metadata.FileSavePathTemplate, _metadata.FileExtension);
             return _fileStorageService.SaveFileAsync(_metadata.FileFolderName, fileName, packageFile, overwrite);
         }
 
         public Task<Stream> DownloadPackageFileAsync(Package package)
         {
-            var fileName = BuildFileName(package, _metadata.FileSavePathTemplate, _metadata.FileExtension);
+            var fileName = FileNameHelper.BuildFileName(package, _metadata.FileSavePathTemplate, _metadata.FileExtension);
             return _fileStorageService.GetFileAsync(_metadata.FileFolderName, fileName);
         }
 
         public Task<Uri> GetPackageReadUriAsync(Package package)
         {
-            var fileName = BuildFileName(package, _metadata.FileSavePathTemplate, _metadata.FileExtension);
+            var fileName = FileNameHelper.BuildFileName(package, _metadata.FileSavePathTemplate, _metadata.FileExtension);
             return _fileStorageService.GetFileReadUriAsync(_metadata.FileFolderName, fileName, endOfAccess: null);
         }
 
         public Task<bool> DoesPackageFileExistAsync(Package package)
         {
-            var fileName = BuildFileName(package, _metadata.FileSavePathTemplate, _metadata.FileExtension);
+            var fileName = FileNameHelper.BuildFileName(package, _metadata.FileSavePathTemplate, _metadata.FileExtension);
             return _fileStorageService.FileExistsAsync(_metadata.FileFolderName, fileName);
         }
 
@@ -65,7 +63,7 @@ namespace NuGetGallery
                 throw new ArgumentNullException(nameof(packageFile));
             }
 
-            var fileName = BuildFileName(
+            var fileName = FileNameHelper.BuildFileName(
                 package,
                 _metadata.FileSavePathTemplate,
                 _metadata.FileExtension);
@@ -79,7 +77,7 @@ namespace NuGetGallery
 
         public Task<Stream> DownloadValidationPackageFileAsync(Package package)
         {
-            var fileName = BuildFileName(
+            var fileName = FileNameHelper.BuildFileName(
                 package,
                 _metadata.FileSavePathTemplate,
                 _metadata.FileExtension);
@@ -95,7 +93,7 @@ namespace NuGetGallery
             }
 
             var normalizedVersion = NuGetVersionFormatter.Normalize(version);
-            var fileName = BuildFileName(
+            var fileName = FileNameHelper.BuildFileName(
                 id,
                 normalizedVersion,
                 _metadata.FileSavePathTemplate,
@@ -118,7 +116,7 @@ namespace NuGetGallery
 
             var normalizedVersion = NuGetVersionFormatter.Normalize(version);
 
-            var fileName = BuildFileName(id, normalizedVersion, _metadata.FileSavePathTemplate, _metadata.FileExtension);
+            var fileName = FileNameHelper.BuildFileName(id, normalizedVersion, _metadata.FileSavePathTemplate, _metadata.FileExtension);
             return _fileStorageService.DeleteFileAsync(_metadata.FileFolderName, fileName);
         }
 
@@ -126,7 +124,7 @@ namespace NuGetGallery
         {
             package = package ?? throw new ArgumentNullException(nameof(package));
 
-            var fileName = BuildFileName(
+            var fileName = FileNameHelper.BuildFileName(
                 package,
                 _metadata.FileSavePathTemplate,
                 _metadata.FileExtension);
@@ -136,7 +134,7 @@ namespace NuGetGallery
 
         public Task<bool> DoesValidationPackageFileExistAsync(Package package)
         {
-            var fileName = BuildFileName(package, _metadata.FileSavePathTemplate, _metadata.FileExtension);
+            var fileName = FileNameHelper.BuildFileName(package, _metadata.FileSavePathTemplate, _metadata.FileExtension);
             return _fileStorageService.FileExistsAsync(_metadata.ValidationFolderName, fileName);
         }
 
@@ -201,76 +199,6 @@ namespace NuGetGallery
             }
         }
 
-        public Task SaveLicenseFileAsync(Package package, Stream licenseFile)
-        {
-            if (package == null)
-            {
-                throw new ArgumentNullException(nameof(package));
-            }
-
-            if (licenseFile == null)
-            {
-                throw new ArgumentNullException(nameof(licenseFile));
-            }
-
-            if (package.EmbeddedLicenseType == EmbeddedLicenseFileType.Absent)
-            {
-                throw new ArgumentException("Package must have an embedded license", nameof(package));
-            }
-
-            var fileName = BuildLicenseFileName(package);
-
-            // Gallery will generally ignore the content type on license files and will use the value from the DB,
-            // but we'll be nice and try to specify correct content type for them.
-            var contentType = package.EmbeddedLicenseType == EmbeddedLicenseFileType.Markdown
-                ? CoreConstants.MarkdownContentType
-                : CoreConstants.TextContentType;
-
-            return _fileStorageService.SaveFileAsync(_metadata.PackageContentFolderName, fileName, contentType, licenseFile, overwrite: true);
-        }
-
-        public Task<Stream> DownloadLicenseFileAsync(Package package)
-        {
-            var fileName = BuildLicenseFileName(package);
-            return _fileStorageService.GetFileAsync(_metadata.PackageContentFolderName, fileName);
-        }
-
-        public Task DeleteLicenseFileAsync(string id, string version)
-        {
-            if (id == null)
-            {
-                throw new ArgumentNullException(nameof(id));
-            }
-
-            if (string.IsNullOrWhiteSpace(id))
-            {
-                throw new ArgumentException($"{nameof(id)} cannot be empty", nameof(id));
-            }
-
-            if (version == null)
-            {
-                throw new ArgumentNullException(nameof(version));
-            }
-
-            if (string.IsNullOrWhiteSpace(version))
-            {
-                throw new ArgumentException($"{nameof(version)} cannot be empty", nameof(version));
-            }
-
-            var normalizedVersion = NuGetVersionFormatter.Normalize(version);
-            var fileName = BuildLicenseFileName(id, normalizedVersion);
-
-            return _fileStorageService.DeleteFileAsync(_metadata.PackageContentFolderName, fileName);
-        }
-
-        private string LicensePathTemplate => $"{_metadata.PackageContentPathTemplate}/{LicenseFileName}";
-
-        private string BuildLicenseFileName(Package package)
-            => BuildFileName(package, LicensePathTemplate, string.Empty);
-
-        private string BuildLicenseFileName(string id, string version)
-            => BuildFileName(id, version, LicensePathTemplate, string.Empty);
-
         private static string BuildBackupFileName(string id, string version, string hash, string extension, string fileBackupSavePathTemplate)
         {
             if (id == null)
@@ -296,52 +224,6 @@ namespace NuGetGallery
                 id.ToLowerInvariant(),
                 version.ToLowerInvariant(),
                 HttpServerUtility.UrlTokenEncode(hashBytes),
-                extension);
-        }
-
-        protected static string BuildFileName(Package package, string format, string extension)
-        {
-            if (package == null)
-            {
-                throw new ArgumentNullException(nameof(package));
-            }
-
-            if (package.PackageRegistration == null ||
-                String.IsNullOrWhiteSpace(package.PackageRegistration.Id) ||
-                (String.IsNullOrWhiteSpace(package.NormalizedVersion) && String.IsNullOrWhiteSpace(package.Version)))
-            {
-                throw new ArgumentException(CoreStrings.PackageIsMissingRequiredData, nameof(package));
-            }
-
-            return BuildFileName(
-                package.PackageRegistration.Id,
-                string.IsNullOrEmpty(package.NormalizedVersion) ?
-                    NuGetVersionFormatter.Normalize(package.Version) :
-                    package.NormalizedVersion, format, extension);
-        }
-
-        protected static string BuildFileName(string id, string version, string pathTemplate, string extension)
-        {
-            if (id == null)
-            {
-                throw new ArgumentNullException(nameof(id));
-            }
-
-            if (version == null)
-            {
-                throw new ArgumentNullException(nameof(version));
-            }
-
-            // Note: packages should be saved and retrieved in blob storage using the lower case version of their filename because
-            // a) package IDs can and did change case over time
-            // b) blob storage is case sensitive
-            // c) we don't want to hit the database just to look up the right case
-            // and remember - version can contain letters too.
-            return String.Format(
-                CultureInfo.InvariantCulture,
-                pathTemplate,
-                id.ToLowerInvariant(),
-                version.ToLowerInvariant(),
                 extension);
         }
     }
