+++ src/NuGetGallery.Core/Services/ICorePackageFileService.cs
@@ -21,19 +21,9 @@ namespace NuGetGallery
         Task SavePackageFileAsync(Package package, Stream packageFile, bool overwrite);
 
         /// <summary>
-        /// Saves the license file to the public container for package content.
-        /// </summary>
-        Task SaveLicenseFileAsync(Package package, Stream licenseFile);
-
-        /// <summary>
         /// Downloads the package from the file storage and reads it into a stream.
         /// </summary>
         Task<Stream> DownloadPackageFileAsync(Package package);
-
-        /// <summary>
-        /// Downloads previously saved license file for a specified package.
-        /// </summary>
-        Task<Stream> DownloadLicenseFileAsync(Package package);
 
         /// <summary>
         /// Generates the URL for the specified package in the public container for available packages.
@@ -97,14 +87,6 @@ namespace NuGetGallery
         /// <param name="id">The package ID. This value is case-insensitive.</param>
         /// <param name="version">The package version. This value is case-insensitive and need not be normalized.</param>
         Task DeletePackageFileAsync(string id, string version);
-
-        /// <summary>
-        /// Deletes the license file for the package from the publicly available storage for the package content.
-        /// </summary>
-        /// <param name="id">The package ID. This value is case-insensitive.</param>
-        /// <param name="version">The package version. This value is case-insensitive and need not be normalized.</param>
-        /// <returns></returns>
-        Task DeleteLicenseFileAsync(string id, string version);
 
         /// <summary>
         /// Copies the contents of the package represented by the stream into the file storage backup location.
